Rename face detection modules
