+++ terran/face/detection/retinaface/__init__.py
@@ -1,0 +1,1 @@
+from terran.face.detection.retinaface.wrapper import RetinaFace  # noqa
